ai_service/requirements.txt
torch
transformers
torch
Pillow
psycopg2-binary
python-dotenv
pinecone
sentence-transformers
